Extract AddAliasDialog.ts
